fix(am-dbg): add fixes for large amount of clients
- fix(am-dbg): fix log scrolling H
- fix(am-dbg): optimized client list >1k
- fix(am-dbg): fix races during GC
