+++ tools/debugger/debugger.go
@@ -8,7 +8,6 @@ import (
 	"encoding/gob"
 	"fmt"
 	"log"
-	"math"
 	"net/http"
 	"net/url"
 	"os"
@@ -46,7 +45,7 @@ const (
 	// TODO customize
 	playInterval = 500 * time.Millisecond
 	// TODO add param --max-clients
-	maxClients            = 1000
+	maxClients            = 5000
 	timeFormat            = "15:04:05.000000000"
 	fastJumpAmount        = 50
 	arrowThrottleMs       = 200
@@ -375,7 +374,8 @@ type Debugger struct {
 	lastScrolledTxTime time.Time
 	repaintScheduled   atomic.Bool
 	// update client list scheduled
-	updateMLScheduled  atomic.Bool
+	updateCLScheduled  atomic.Bool
+	buildCLScheduled   atomic.Bool
 	lastKeystroke      tcell.Key
 	lastKeystrokeTime  time.Time
 	updateLogScheduled atomic.Bool
@@ -397,7 +397,9 @@ type Debugger struct {
 	tagsBar         *cview.TextView
 	clip            clipper.Clipboard
 	// toolbarItems is a list of row of toolbars items
-	toolbarItems [][]toolbarItem
+	toolbarItems     [][]toolbarItem
+	msgsDelayed      []*telemetry.DbgMsgTx
+	msgsDelayedConns []string
 }
 
 // New creates a new debugger instance and optionally import a data file.
@@ -425,7 +427,7 @@ func New(ctx context.Context, opts Opts) (*Debugger, error) {
 	gob.Register(Exportable{})
 	gob.Register(am.Relation(0))
 
-	id := utils.RandID(0)
+	id := utils.RandId(0)
 	if opts.ID != "" {
 		id = opts.ID
 	}
@@ -921,6 +923,7 @@ func (d *Debugger) updateAddressBar() {
 	if d.C != nil {
 		machId = d.C.id
 		if d.C.CursorTx1 > 0 {
+			// TODO conflict with GC?
 			txId = d.C.MsgTxs[d.C.CursorTx1-1].ID
 		}
 		machConn = d.C.connected.Load()
@@ -1315,279 +1318,6 @@ func (d *Debugger) updateTimelines() {
 	d.timelineSteps.SetEmptyRune(' ')
 }
 
-func (d *Debugger) updateClientList(immediate bool) {
-	if immediate {
-		d.doUpdateClientList(true)
-		return
-	}
-
-	if !d.updateMLScheduled.CompareAndSwap(false, true) {
-		// debounce non-forced updates
-		return
-	}
-
-	go func() {
-		time.Sleep(sidebarUpdateDebounce)
-		d.doUpdateClientList(false)
-	}()
-}
-
-// TODO rewrite, update via a state
-func (d *Debugger) doUpdateClientList(immediate bool) {
-	if d.Mach.IsDisposed() || d.Mach.Is1(ss.SelectingClient) ||
-		d.Mach.Is1(ss.HelpDialog) {
-		return
-	}
-
-	update := func() {
-		_, _, width, _ := d.clientList.GetRect()
-		maxLen := width - 13
-		if maxLen < 5 {
-			maxLen = 15
-		}
-
-		// count
-		longestName := 0
-		for _, item := range d.clientList.GetItems() {
-			ref := item.GetReference().(*sidebarRef)
-			l := len(ref.name) + ref.lvl
-			if l > longestName {
-				longestName = l
-			}
-		}
-		if longestName > maxLen {
-			longestName = maxLen
-		}
-
-		// update
-		for i, item := range d.clientList.GetItems() {
-			ref := item.GetReference().(*sidebarRef)
-			c := d.Clients[ref.name]
-
-			spacePre := ""
-			spacePost := " "
-			if d.clientHasParent(c.id) {
-				spacePre = " "
-			}
-
-			name := strings.Repeat("-", ref.lvl) + spacePre + ref.name
-			if len(name) > maxLen {
-				name = name[:maxLen-2] + ".."
-			}
-
-			spaceCount := int(math.Max(0, float64(longestName+1-len(name))))
-			namePad := name +
-				strings.Repeat(" ", spaceCount) + spacePost
-			label := d.getClientListLabel(namePad, c, i)
-			item.SetMainText(label)
-		}
-
-		if len(d.Clients) > 0 {
-			var totalSum uint64
-			for _, c := range d.Clients {
-				totalSum += c.mTimeSum
-			}
-
-			d.clientList.SetTitle(d.P.Sprintf(
-				" Machines:%d T:%v ", len(d.Clients), totalSum))
-		} else {
-			d.clientList.SetTitle(" Machines ")
-		}
-
-		// render if delayed
-		if !immediate {
-			d.draw(d.clientList)
-		}
-	}
-
-	// avoid eval in handlers
-	if immediate {
-		update()
-	} else {
-		d.Mach.Eval("doUpdateClientList", update, nil)
-	}
-
-	d.updateMLScheduled.Store(false)
-}
-
-// buildClientList builds the clientList with the list of clients.
-// selectedIndex: index of the selected item, -1 for the current one.
-func (d *Debugger) buildClientList(selectedIndex int) {
-	// prev state
-	selected := ""
-	var item *cview.ListItem
-	if selectedIndex == -1 {
-		item = d.clientList.GetCurrentItem()
-	} else if selectedIndex > -1 {
-		item = d.clientList.GetItem(selectedIndex)
-	}
-	if item != nil {
-		selected = item.GetReference().(*sidebarRef).name
-	}
-
-	// re-gen all
-	d.clientList.Clear()
-	var list []string
-	for _, c := range d.Clients {
-		list = append(list, c.id)
-	}
-
-	// sort a-z, with value numbers
-	humanSort(list)
-
-	pos := 0
-	// TODO REWRITE
-	for _, parent := range list {
-		if d.clientHasParent(parent) {
-			continue
-		}
-
-		// create list item
-		item := cview.NewListItem(parent)
-		item.SetReference(&sidebarRef{name: parent, lvl: 0})
-		d.clientList.AddItem(item)
-
-		if selected == "" && d.C != nil && d.C.id == parent {
-			// pre-select the current machine
-			d.clientList.SetCurrentItem(pos)
-		} else if selected == parent {
-			// select the previously selected one
-			d.clientList.SetCurrentItem(pos)
-		}
-
-		pos = d.clientListChild(list, parent, pos, selected, 1)
-
-		pos++
-	}
-
-	var totalSum uint64
-	for _, c := range d.Clients {
-		totalSum += c.mTimeSum
-	}
-
-	d.clientList.SetTitle(d.P.Sprintf(
-		" Machines:%d T:%v ", len(d.Clients), totalSum))
-
-	// sort TODO blinks anyway
-	d.doUpdateClientList(true)
-}
-
-type sidebarRef struct {
-	name string
-	lvl  int
-}
-
-func (d *Debugger) clientListChild(
-	list []string, parent string, pos int, selected string,
-	lvl int,
-) int {
-	for _, child := range list {
-		cc := d.Clients[child]
-
-		if !d.clientHasParent(child) || cc.MsgStruct.Parent != parent {
-			continue
-		}
-
-		// create list item
-		item := cview.NewListItem(child)
-		item.SetReference(&sidebarRef{name: child, lvl: lvl})
-		d.clientList.AddItem(item)
-
-		if selected == "" && d.C != nil && d.C.id == child {
-			// pre-select the current machine
-			d.clientList.SetCurrentItem(pos)
-		} else if selected == child {
-			// select the previously selected one
-			d.clientList.SetCurrentItem(pos)
-		}
-
-		pos = d.clientListChild(list, child, pos, selected, lvl+1)
-
-		pos++
-	}
-	return pos
-}
-
-func (d *Debugger) clientHasParent(cid string) bool {
-	c, ok := d.Clients[cid]
-	if !ok || c.MsgStruct.Parent == "" {
-		return false
-	}
-	_, ok = d.Clients[c.MsgStruct.Parent]
-
-	return ok
-}
-
-func (d *Debugger) getClientListLabel(
-	name string, c *Client, index int,
-) string {
-	isHovered := d.clientList.GetCurrentItemIndex() == index
-	hasFocus := d.Mach.Is1(ss.SidebarFocused)
-
-	var currCTxIdx int
-	var currCTx *telemetry.DbgMsgTx
-	// current tx of the selected client
-	currSelTx := d.currentTx()
-	if currSelTx != nil {
-		currTime := d.lastScrolledTxTime
-		if currTime.IsZero() {
-			currTime = *currSelTx.Time
-		}
-		currCTxIdx = c.lastTxTill(currTime)
-		if currCTxIdx != -1 {
-			currCTx = c.MsgTxs[currCTxIdx]
-		}
-	}
-
-	// Ready, Start, Error states indicators
-	var state string
-	isErrNow := false
-	if currCTx != nil {
-		readyIdx := slices.Index(c.MsgStruct.StatesIndex, ssam.BasicStates.Ready)
-		startIdx := slices.Index(c.MsgStruct.StatesIndex, ssam.BasicStates.Start)
-		errIdx := slices.Index(c.MsgStruct.StatesIndex, am.Exception)
-		isErrNow = errIdx != -1 && am.IsActiveTick(currCTx.Clocks[errIdx])
-		if readyIdx != -1 && am.IsActiveTick(currCTx.Clocks[readyIdx]) {
-			state = "R"
-		} else if startIdx != -1 && am.IsActiveTick(currCTx.Clocks[startIdx]) {
-			state = "S"
-		}
-		// push to the front
-		if isErrNow {
-			state = "E" + state
-		}
-	}
-	if state == "" {
-		state = " "
-	}
-
-	label := d.P.Sprintf("%s %s|%d", name, state, currCTxIdx+1)
-	if currCTxIdx+1 < len(c.MsgTxs) {
-		label += "+"
-	}
-
-	// mark selected
-	if d.C != nil && c.id == d.C.id {
-		label = "[::bu]" + label
-	}
-
-	if isErrNow {
-		label = "[red]" + label
-	} else if c.hadErrSinceTx(currCTxIdx, 100) {
-		label = "[orangered]" + label
-	} else if !c.connected.Load() {
-		if isHovered && !hasFocus {
-			label = "[grey]" + label
-		} else if !isHovered {
-			label = "[grey]" + label
-		} else {
-			label = "[black]" + label
-		}
-	}
-
-	return label
-}
-
 func (d *Debugger) updateBorderColor() {
 	color := colorInactive
 	if d.Mach.IsErr() {
@@ -1726,6 +1456,10 @@ func (d *Debugger) doCleanOnConnect() bool {
 }
 
 func (d *Debugger) updateMatrix() {
+	if !d.Mach.Any1(ss.MatrixView, ss.TreeMatrixView) {
+		return
+	}
+
 	if d.Mach.Is1(ss.MatrixRain) {
 		d.updateMatrixRain()
 	} else {
@@ -2085,6 +1819,7 @@ func (d *Debugger) getSidebarCurrClientIdx() int {
 // filterClientTxs toolbarItems client's txs according the selected
 // toolbarItems. Called by toolbarItem states, not directly.
 func (d *Debugger) filterClientTxs() {
+	// TODO remove with SQL
 	if d.C == nil || !d.isFiltered() {
 		return
 	}
@@ -2214,7 +1949,7 @@ func (d *Debugger) ProcessFilterChange(ctx context.Context, filterTxs bool) {
 	d.updateClientList(false)
 	d.updateToolbar()
 	d.updateTimelines()
-	d.updateMatrixRain()
+	d.updateMatrix()
 	d.updateLog(false)
 	d.draw()
 }
